add binary shared object for windows

+++ install.js
@@ -2,9 +2,14 @@ var fs = require("fs");
 var path = require('path');
 var parentPath = path.dirname(__filename);
 var nativeModule = parentPath + "/binary/" + process.platform + "_" + process.arch + "/";
-
+var devMode = true;
 var isNative = false;
-if (fs.existsSync(nativeModule) && !fs.existsSync("Makefile")) {
+
+if (!fs.existsSync("Makefile")) {
+    devMode = false;
+}
+
+if (fs.existsSync(nativeModule) && !devMode) {
     try {
         var PL = require(nativeModule +'phplike' );
         var res = PL.exec("echo a");
@@ -20,7 +25,8 @@ if (fs.existsSync(nativeModule) && !fs.existsSync("Makefile")) {
 if (isNative == false) {
     var fileHandle = require('./nodejs/file.js');
     var isForce = true;
-    fileHandle.rmdir('./binary', isForce);
+
+    if (!devMode) fileHandle.rmdir('./binary', isForce);
     console.log("Auto Recompile C/C++ library.");
     var proce = require('child_process');
